sprint 23

+++ app/src/main/java/com/example/playlistmarker/data/db/sharedpreferences/AddedAtStorage.kt
@@ -5,15 +5,15 @@ import android.content.SharedPreferences
 class AddedAtStorage(private val sharedPreferences: SharedPreferences) {
 
 
-    fun saveAddedTime(trackId: Int, time: Long = System.currentTimeMillis()) {
+    fun saveAddedTime(trackId: Long, time: Long = System.currentTimeMillis()) {
         sharedPreferences.edit().putLong("added_at_$trackId", time).apply()
     }
 
-    fun getAddedTime(trackId: Int): Long {
+    fun getAddedTime(trackId: Long): Long {
         return sharedPreferences.getLong("added_at_$trackId", 0L)
     }
 
-    fun removeAddedTime(trackId: Int) {
+    fun removeAddedTime(trackId: Long) {
         sharedPreferences.edit().remove("added_at_$trackId").apply()
     }
 }
